feat(swap): refactor swapselector
- separates component + styling logic into separate files

- renamed components to better reflect their purpose
